feat: new k8s request duration histogram metric
This is a new metric which is a superset of the k8s_request_total
metric. The old metric is retained for backwards compatibility.

A common issue with workflows is that it stresses the kubernetes
API. Using this metric you can get insights into how long each type of
request is taking which should enable optimizations to help alleviate
problems in this area.

The new metric has the same attributes/labels of `kind`, `verb` and
`status_code` but now emits them with a time duration for each type of
response.

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>