feat: new cron workflow trigger counter metric
From #12589.

A new metric which counts how many times each cron workflow has
triggered. A simple enough counter which can be checked against
expectations for the cron.

Note to reviewers: this is part of a stack of reviews for metrics
changes. Please don't merge until the rest of the stack is also ready.

Signed-off-by: Alan Clucas <alan@clucas.org>